+++ packages/gateway/src/commands/subgraph.ts
@@ -147,6 +147,12 @@ export const addCommand: AddCommand = (ctx, cli) =>
           return [...builtinPlugins, ...userPlugins];
         },
       };
+      config.renderLegacyGraphiQL ||= opts.renderLegacyGraphiql;
+      if (config.renderLegacyGraphiQL) {
+        // Setting these to undefined will make the runtime use default GraphiQL
+        config.renderGraphiQL = undefined;
+        config.playgroundName = undefined;
+      }
       if (hivePersistedDocumentsEndpoint) {
         const token =
           hivePersistedDocumentsToken ||