+++ vitest.config.ts
@@ -61,6 +61,7 @@ export default defineConfig({
                   '!**/e2e/hoist-and-prefix-transform', // has transforms
                   '!**/e2e/cloudflare-workers', // runs gateway in proxy mode
                   '!**/e2e/edfs-mesh-defs-with-entity-resolution', // entity resolution through additional resolvers subscriptions
+                  '!**/e2e/typed-sdk', // nothing to do with the runtime
                 ]
               : []),
           ],