+++ packages/plugins/prometheus/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @graphql-mesh/plugin-prometheus
 
+## 2.1.41
+### Patch Changes
+
+- Updated dependencies []:
+  - @graphql-hive/gateway-runtime@2.8.11
+
 ## 2.1.40
 ### Patch Changes